port check 004: check_DESCRIPTION_is_propper_HTML_snippet
(issue #1388)
