+++ Lib/fontbakery/checks.py
@@ -122,37 +122,6 @@ def check_file_is_named_canonically(fb, font_fname):
     return False
 
 
-def check_DESCRIPTION_is_propper_HTML_snippet(fb, descfile):
-  """When packaging families for google/fonts. If there is no
-  DESCRIPTION.en_us.html file, the add_font.py metageneration tool will
-  insert a dummy description file which contains invalid html.
-  This file needs to either be replaced with an existing description file
-  or edited by hand."""
-  fb.new_check("004", "Is this a propper HTML snippet ?")
-  try:
-    import magic
-    contenttype = magic.from_file(descfile)
-    if "HTML" not in contenttype:
-      data = open(descfile).read()
-      if "<p>" in data and "</p>" in data:
-        fb.ok(("{} is a propper"
-               " HTML snippet.").format(descfile))
-      else:
-        fb.error(("{} is not a propper"
-                  " HTML snippet.").format(descfile))
-    else:
-      fb.ok("{} is a propper HTML file.".format(descfile))
-  except AttributeError:
-     fb.skip("python magic version mismatch: "
-             "This check was skipped because the API of the python"
-             " magic module version installed in your system does not"
-             " provide the from_file method used in"
-             " the check implementation.")
-  except ImportError:
-     fb.skip("This check depends on the magic python module which"
-             " does not seem to be currently installed on your system.")
-
-
 def check_font_designer_field_is_not_unknown(fb, family):
   fb.new_check("007", "Font designer field is 'unknown' ?")
   if family.designer.lower() == 'unknown':
